ChatSIEM/requirements.txt
# Core Dependencies
elasticsearch==8.11.0
requests==2.31.0
python-dotenv==1.0.0

# NLP and AI
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
spacy==3.7.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# API Framework
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1